+++ package.json
@@ -14,8 +14,8 @@
   },
   "main": "gofor/index.js",
   "scripts": {
-    "test": "mocha test.js **/test.js --inspect --require configuration/test_setup.js",
-    "test:single": "mocha --inspect --require configuration/test_setup.js",
+    "test": "mocha --opts configuration/mocha.opts **/test.js **/*.test.js",
+    "test:single": "mocha --opts configuration/mocha.opts ",
     "lint": "eslint -c .eslintrc '*.js' '**/*.js'",
     "prepublishOnly": "npm t && npm run lint"
   },
